Refactor file structure for comparison operations
