+++ src/tensor_ops/mod.rs
@@ -185,6 +185,7 @@ pub use add::{add, TryAdd};
 pub use bce::bce_with_logits;
 pub use broadcast_to::BroadcastTo;
 pub use clamp::clamp;
+pub use cmp::eq;
 pub use cos::cos;
 pub use div::{div, TryDiv};
 pub use dropout::dropout;
